+++ src/main/java/de/samply/query/Query.java
@@ -1,7 +1,11 @@
 package de.samply.query;
 
+import de.samply.converter.Format;
+
 public record Query(
     String queryId,
     String query,
-    QueryFormat queryFormat
-) {}
+    Format queryFormat
+) {
+
+}