+++ src/main/java/de/samply/query/QueryFormat.java
@@ -1,6 +1,0 @@
-package de.samply.query;
-
-public enum QueryFormat {
-  FHIR,
-  CQL
-}