+++ src/main/java/de/samply/result/ResultFormat.java
@@ -1,8 +1,0 @@
-package de.samply.result;
-
-public enum ResultFormat {
-  CSV,
-  EXCEL,
-  JAVA_OBJECT,
-  FHIR
-}