+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - FHIR Rev Includes in Containers Templates
 - Executor in Fhir Store Client
 - Converter Pattern
+- Converter Manager